chore: cleanup layout setup
